requirements.txt
psycopg2-binary
bcrypt
python-dotenv
Flask==3.0.3
gunicorn==21.2.0